Adding 2 base cases
