+++ edge-functions/streaming-response/pages/api/01-simple.ts
@@ -1,0 +1,21 @@
+import { NextResponse, NextRequest } from 'next/server'
+const decoder = new TextDecoder()
+
+export const config = {
+  runtime: 'experimental-edge',
+}
+
+const stream = new ReadableStream({
+  start(controller) {
+    controller.enqueue(
+      '<html><head><title>Vercel Edge Functions + Streaming</title></head><body>'
+    )
+    controller.enqueue('Vercel Edge Functions + Streaming')
+    controller.enqueue('</body></html>')
+    controller.close()
+  },
+})
+
+export default async function handler(req: NextRequest, res: NextResponse) {
+  return new Response(stream, { headers: { 'Content-Type': 'text/html' } })
+}
